Not a full commit
Just some of the requested changes for Shiela:
- Moved Article back to models.py
- put setting.py back to original
- put rmc.css in static/css/ (link not yet changed in html this commit)
