+++ companionpages/companionpages/urls.py
@@ -1,6 +1,10 @@
 from django.conf.urls import patterns, include, url
-from django.contrib import admin
 
+import autocomplete_light
+#autocomplete_light.registry.autocomplete_model_base = YourAutocompleteModelBase
+autocomplete_light.autodiscover()
+
+from django.contrib import admin
 admin.autodiscover()
 
 from envelope.views import ContactView
@@ -23,6 +27,8 @@ urlpatterns = patterns('',
     url(r'^admin/doc/', include('django.contrib.admindocs.urls')),
     url(r'^admin/', include(admin.site.urls)),
     url(r'^companionpages', include('supportingmaterials.urls')),
+    url(r'^autocomplete/', include('autocomplete_light.urls')),
+
 )
 
 urlpatterns += patterns('django.contrib.flatpages.views',
