+++ companionpages/home/views.py
@@ -5,7 +5,7 @@ from django.views.generic import TemplateView
 
 from .models import Faq
 from news.models import News
-from supportingmaterials.models import Article, CompanionForm
+from supportingmaterials.models import Article
 
 class HomeView(TemplateView):
     template_name="base.html"
@@ -29,4 +29,4 @@ class AboutView(TemplateView):
 #class CompanionView(TemplateView):
 #	form = CompanionForm(extra=request.POST.get('extra_field_count'))
 #	template_name='create.html'
-	#return render(request, "template", { 'form': form })
+#	return render(request, "template", { 'form': form })
