+++ companionpages/static/css/rmc.css
@@ -1,0 +1,119 @@
+#subheader {
+    text-align: center;
+    color: orange;
+    padding:0px;
+    display:none;
+}
+#subheader li{
+    display:inline;
+    margin-right:15px;
+    box-shadow: 0 0 0 0; 
+}
+#subheader li:hover{
+    display:inline;
+    margin-right:15px;
+    color:red;
+}
+.red {color:red}
+p{ line-height: 1em; }
+body{
+    font-family: arial; font-size: 80%; 
+    line-height: 1.4em; width: 100%; height:100%;
+    margin: 0; background: #eee;
+    min-height:100%;position: absolute;font-size:12px;
+}
+div.tabbed-menu {
+    font-size: 14px;
+    width: 100%;
+}
+
+ul.tabs {
+    text-align: center;
+    list-style: none;
+    padding-bottom: 0px;
+    line-height: 26px;
+    color: #82FA58; /*#0088CC;*/
+    border-bottom: 1px solid #DDDDDD;
+}
+
+
+input {
+border-radius: 4px;
+}
+
+ul.tabs li {
+    /*float: left; */
+    position: relative;
+    margin-bottom: -1px;
+    padding: 3px 10px 0 10px;
+    background: #777;/*#EFEFEF;*/
+    display: inline-block;
+    /*font-family: Helvetica, Arial, sans-serif;    */
+    border-top-left-radius: 6px; 
+    border-top-right-radius: 6px;
+    -webkit-box-shadow: 2px -1px 2px rgba(0, 0, 0, 0.6);
+    -moz-box-shadow: 2px -1px 2px rgba(0, 0, 0, 0.6);
+    -o-box-shadow: 2px -1px 2px rgba(0, 0, 0, 0.6);
+    -ms-box-shadow: 2px -1px 2px rgba(0, 0, 0, 0.6);
+    -khtml-box-shadow: 2px -1px 2px rgba(0, 0, 0, 0.6);
+    box-shadow: 2px -1px 2px rgba(0, 0, 0, 0.6);
+}
+ul.tabs a {
+    text-decoration:none;
+    color: #82FA58;
+}
+.input-tips{
+border-radius: 6px; 
+}
+.linkso a{
+    display:inline-block;
+    text-align: center;
+    border-radius: 6px; 
+    -webkit-box-shadow: 2px -1px 5px rgba(0, 0, 0, 0.6);
+    -moz-box-shadow: 2px -1px 5px rgba(0, 0, 0, 0.6);
+    -ms-box-shadow: 2px -1px 5px rgba(0, 0, 0, 0.6);
+    -o-box-shadow: 2px -1px 5px rgba(0, 0, 0, 0.6);
+    -khtml-box-shadow: 2px -1px 5px rgba(0, 0, 0, 0.6);
+    box-shadow: 2px -1px 5px rgba(0, 0, 0, 0.6);
+    padding: 3px 8px 3px 8px;
+    text-decoration:none;
+    line-height:18px;
+    margin-bottom:5px;
+}
+.linkso{
+    text-align: center;
+}
+.linkso a:hover{
+    color:red;
+}
+ul.tabs li.selected {
+    margin-bottom: -2px;
+    background: #EFEFEF;
+    color: #585858;
+    border-bottom: 1px solid #EFEFEF;
+    -webkit-box-shadow: 2px -3px px rgba(0, 0, 0, 0.6);
+    -moz-box-shadow: 2px -3px 2px rgba(0, 0, 0, 0.6);
+    -o-box-shadow: 2px -3px 2px rgba(0, 0, 0, 0.6);
+    -ms-box-shadow: 2px -3px 2px rgba(0, 0, 0, 0.6);
+    -khtml-box-shadow: 2px -3px 2px rgba(0, 0, 0, 0.6);
+    box-shadow: 2px -3px 2px rgba(0, 0, 0, 0.6);
+    z-index: 4;
+}
+ul.tabs li.selected a {
+color: #585858;
+}
+
+ul.tabs li:hover {
+    color: #333333;
+    cursor: pointer;
+    background: #FFFFFF;
+}
+ul.tabs li:hover a {
+color: #333333;
+}
+
+.anchor {
+top:0;
+position:fixed;
+}
+.page{ width:100%}
